line item embeding into cart at service layer

+++ bookstore-backend/src/main/java/com/bookstore/backend/entities/CartLineItemEntity.java
@@ -7,7 +7,7 @@ import lombok.Data;
  */
 @Data
 public class CartLineItemEntity {
-    private Long id;         // PK Auto-increment
+    private int id;         // PK Auto-increment
     private String userId;   // Associates item to a specific session/user
     private Integer itemId;  // FK pointing to book table
     private int quantity;
